Updated to dev version
Updated  the source of imgareaselect.min.js to source of imgareaselect.dev.js from the github repository: https://github.com/odyniec/imgareaselect/blob/master/jquery.imgareaselect.dev.js to accomodate the update from jquery 1.3.2 to jquery 1.9. See the note on the jquery version update for the reason for this accomodation.